Removing extras...
