--- nlglib/_resources/log.config.yaml
@@ -1,44 +1,0 @@
----
-version: 1
-disable_existing_loggers: False
-formatters:
-    simple:
-        format: "%(name)-20s - %(levelname)-8s - %(message)s"
-    verbose:
-        format: "%(asctime)s %(name)-20s %(levelname)-8s %(threadName)-10s %(funcName)-20s (%(lineno)d):\n\t%(message)s"
-
-handlers:
-    console:
-        class: logging.StreamHandler
-        level: DEBUG
-        formatter: simple
-        stream: ext://sys.stdout
-
-    single_run_handler:
-        class: logging.FileHandler
-        level: DEBUG
-        formatter: simple
-        filename: last_run.log
-        mode: w
-        encoding: utf8
-
-    rotating_lib_handler:
-        class: logging.handlers.RotatingFileHandler
-        level: WARNING
-        formatter: verbose
-        filename: libnlg.log
-        maxBytes: 10485760 # 10MB
-        backupCount: 2
-        encoding: utf8
-
-root:
-    level: DEBUG
-    handlers: [console, single_run_handler, rotating_lib_handler]
-
-loggers:
-    nlg:
-        level: DEBUG
-        handlers: [console, single_run_handler, rotating_lib_handler]
-        propagate: no
-
-...
