--- nlglib/_resources/simplenlg.settings
@@ -1,3 +1,0 @@
-SimplenlgHost     localhost
-SimplenlgPort     50007
-SimplenlgJarPath  nlglib/resources/simplenlg.jar
