--- nlglib/lexicon/__init__.py
@@ -1,1 +1,0 @@
-from .default import Lexicon as DefaultLexicon
