--- nlglib/signals.py
@@ -1,53 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    nlglib.signals
-    ~~~~~~~~~~~~~~
-
-    Implements signals based on blinker if available, otherwise
-    falls silently back to a noop.
-
-    :copyright: (c) 2015 by Armin Ronacher.
-    :license: BSD, see LICENSE for more details.
-"""
-signals_available = False
-try:
-    from blinker import Namespace
-    signals_available = True
-except ImportError:
-    class Namespace(object):
-        def signal(self, name, doc=None):
-            return _FakeSignal(name, doc)
-
-    class _FakeSignal(object):
-        """If blinker is unavailable, create a fake class with the same
-        interface that allows sending of signals but will fail with an
-        error on anything else.  Instead of doing anything on send, it
-        will just ignore the arguments and do nothing instead.
-        """
-
-        def __init__(self, name, doc=None):
-            self.name = name
-            self.__doc__ = doc
-
-        def _fail(self, *args, **kwargs):
-            raise RuntimeError('signalling support is unavailable '
-                               'because the blinker library is '
-                               'not installed.')
-        send = lambda *a, **kw: None
-        connect = disconnect = has_receivers_for = receivers_for = \
-            temporarily_connected_to = connected_to = _fail
-        del _fail
-
-# The namespace for code signals. If you are not nlglib code,
-# do not put signals in here. Create your own namespace instead.
-_signals = Namespace()
-
-# Core signals.  For usage examples grep the source code or consult
-# the API documentation in docs/api.rst as well as docs/signals.rst
-lexicon_context_pushed = _signals.signal('lexicon_context-pushed')
-lexicon_context_popped = _signals.signal('lexicon_context-popped')
-pipeline_context_pushed = _signals.signal('pipeline_context-pushed')
-pipeline_context_popped = _signals.signal('pipeline_context-popped')
-pipeline_context_tearing_down = _signals.signal('pipeline_context-popped')
-linguistic_context_pushed = _signals.signal('linguistic_context-pushed')
-linguistic_context_popped = _signals.signal('linguistic_context-popped')
